refactor: producer, consumer 분리 및 경로 수정
