+++ src/backend/chat-service/src/main/java/capstone/chatservice/domain/dm/controller/DirectMessageCommandController.java
@@ -7,9 +7,9 @@ import capstone.chatservice.domain.dm.dto.request.DirectMessageModifyRequest;
 import capstone.chatservice.domain.dm.service.command.DirectMessageCommandService;
 import capstone.chatservice.domain.model.UploadFile;
 import capstone.chatservice.infra.S3.FileStore;
-import capstone.chatservice.infra.kafka.producer.KafkaProducer;
-import capstone.chatservice.infra.kafka.producer.alarm.AlarmProducer;
+import capstone.chatservice.infra.kafka.producer.alarm.AlarmEventProducer;
 import capstone.chatservice.infra.kafka.producer.alarm.dto.DmAlarmEventDto;
+import capstone.chatservice.infra.kafka.producer.chat.ChatEventProducer;
 import java.io.IOException;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -24,31 +24,31 @@ import org.springframework.web.multipart.MultipartFile;
 public class DirectMessageCommandController {
 
     private final FileStore fileStore;
-    private final AlarmProducer alarmProducer;
-    private final KafkaProducer kafkaProducer;
+    private final AlarmEventProducer alarmEventProducer;
+    private final ChatEventProducer chatEventProducer;
     private final DirectMessageCommandService commandService;
 
     @MessageMapping("/direct/message/send")
     public void save(DirectMessageCreateRequest createRequest) {
 
         DirectMessageDto messageDto = commandService.save(createRequest);
-        kafkaProducer.sendToDirectChatTopic(messageDto);
+        chatEventProducer.sendToDirectChatTopic(messageDto);
         DmAlarmEventDto dmAlarmEventDto = DmAlarmEventDto.from(createRequest);
-        alarmProducer.sendToDmAlarmEventTopic(dmAlarmEventDto);
+        alarmEventProducer.sendToDmAlarmEventTopic(dmAlarmEventDto);
     }
 
     @MessageMapping("/direct/message/modify")
     public void modify(DirectMessageModifyRequest modifyRequest) {
 
         DirectMessageDto messageDto = commandService.modify(modifyRequest);
-        kafkaProducer.sendToDirectChatTopic(messageDto);
+        chatEventProducer.sendToDirectChatTopic(messageDto);
     }
 
     @MessageMapping("/direct/message/delete")
     public void modify(DirectMessageDeleteRequest deleteRequest) {
 
         DirectMessageDto messageDto = commandService.delete(deleteRequest);
-        kafkaProducer.sendToDirectChatTopic(messageDto);
+        chatEventProducer.sendToDirectChatTopic(messageDto);
     }
 
     @PostMapping("/direct/message/file")
@@ -59,6 +59,6 @@ public class DirectMessageCommandController {
         List<UploadFile> uploadFiles = fileStore.storeFiles(files);
         createRequest.setFiles(uploadFiles);
         DirectMessageDto messageDto = commandService.save(createRequest);
-        kafkaProducer.sendToDirectChatTopic(messageDto);
+        chatEventProducer.sendToDirectChatTopic(messageDto);
     }
 }
