+++ src/backend/chat-service/src/main/java/capstone/chatservice/domain/forum/controller/ForumMessageCommandController.java
@@ -7,7 +7,7 @@ import capstone.chatservice.domain.forum.dto.request.ForumMessageModifyRequest;
 import capstone.chatservice.domain.forum.service.command.ForumMessageCommandService;
 import capstone.chatservice.domain.model.UploadFile;
 import capstone.chatservice.infra.S3.FileStore;
-import capstone.chatservice.infra.kafka.producer.KafkaProducer;
+import capstone.chatservice.infra.kafka.producer.chat.ChatEventProducer;
 import java.io.IOException;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -22,25 +22,25 @@ import org.springframework.web.multipart.MultipartFile;
 public class ForumMessageCommandController {
 
     private final FileStore fileStore;
-    private final KafkaProducer kafkaProducer;
+    private final ChatEventProducer chatEventProducer;
     private final ForumMessageCommandService commandService;
 
     @MessageMapping("/forum/message/send")
     public void save(ForumMessageCreateRequest createRequest) {
         ForumMessageDto messageDto = commandService.save(createRequest);
-        kafkaProducer.sendToForumChatTopic(messageDto);
+        chatEventProducer.sendToForumChatTopic(messageDto);
     }
 
     @MessageMapping("/forum/message/modify")
     public void modify(ForumMessageModifyRequest modifyRequest) {
         ForumMessageDto messageDto = commandService.modify(modifyRequest);
-        kafkaProducer.sendToForumChatTopic(messageDto);
+        chatEventProducer.sendToForumChatTopic(messageDto);
     }
 
     @MessageMapping("/forum/message/delete")
     public void delete(ForumMessageDeleteRequest deleteRequest) {
         ForumMessageDto messageDto = commandService.delete(deleteRequest);
-        kafkaProducer.sendToForumChatTopic(messageDto);
+        chatEventProducer.sendToForumChatTopic(messageDto);
     }
 
     @PostMapping("/forum/message/file")
@@ -51,6 +51,6 @@ public class ForumMessageCommandController {
         List<UploadFile> uploadFiles = fileStore.storeFiles(files);
         createRequest.setFiles(uploadFiles);
         ForumMessageDto messageDto = commandService.save(createRequest);
-        kafkaProducer.sendToForumChatTopic(messageDto);
+        chatEventProducer.sendToForumChatTopic(messageDto);
     }
 }
