+++ src/backend/chat-service/src/main/java/capstone/chatservice/domain/server/controller/ServerMessageQueryController.java
@@ -5,7 +5,7 @@ import capstone.chatservice.domain.server.dto.request.ServerMessageTypingRequest
 import capstone.chatservice.domain.server.service.query.ServerMessageQueryService;
 import capstone.chatservice.global.common.dto.DataResponseDto;
 import capstone.chatservice.global.common.dto.PageResponseDto;
-import capstone.chatservice.infra.kafka.producer.KafkaProducer;
+import capstone.chatservice.infra.kafka.producer.chat.ChatEventProducer;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.messaging.handler.annotation.MessageMapping;
@@ -17,7 +17,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class ServerMessageQueryController {
 
-    private final KafkaProducer producerService;
+    private final ChatEventProducer producerService;
     private final ServerMessageQueryService queryService;
 
     @GetMapping("/server/messages/channel")
