+++ src/backend/chat-service/src/main/java/capstone/chatservice/global/config/kafka/consumer/chat/DirectChatConsumerConfig.java
@@ -1,6 +1,6 @@
-package capstone.chatservice.global.config.kafka.consumer;
+package capstone.chatservice.global.config.kafka.consumer.chat;
 
-import capstone.chatservice.domain.emoji.dto.EmojiDto;
+import capstone.chatservice.domain.dm.dto.DirectMessageDto;
 import java.util.HashMap;
 import java.util.Map;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
@@ -16,16 +16,17 @@ import org.springframework.kafka.support.serializer.JsonDeserializer;
 
 @EnableKafka
 @Configuration
-public class EmojiChatConsumerConfig {
+public class DirectChatConsumerConfig {
 
     @Value("${spring.kafka.bootstrap-servers}")
     private String bootstrapServers;
 
-    @Value("${spring.kafka.consumer.group-id.emoji-chat}")
+    @Value("${spring.kafka.consumer.group-id.direct-chat}")
     private String groupId;
 
+
     @Bean
-    public Map<String, Object> emojiChatConsumerConfiguration() {
+    public Map<String, Object> directChatConsumerConfiguration() {
         Map<String, Object> config = new HashMap<>();
         config.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServers);
         config.put(ConsumerConfig.GROUP_ID_CONFIG, groupId);
@@ -33,17 +34,17 @@ public class EmojiChatConsumerConfig {
     }
 
     @Bean
-    public ConsumerFactory<String, EmojiDto> emojiChatConsumerFactory() {
+    public ConsumerFactory<String, DirectMessageDto> directChatConsumerFactory() {
         return new DefaultKafkaConsumerFactory<>(
-                emojiChatConsumerConfiguration(),
+                directChatConsumerConfiguration(),
                 new StringDeserializer(),
-                new JsonDeserializer<>(EmojiDto.class));
+                new JsonDeserializer<>(DirectMessageDto.class));
     }
 
     @Bean
-    public ConcurrentKafkaListenerContainerFactory<String, EmojiDto> emojiChatListenerContainerFactory() {
-        ConcurrentKafkaListenerContainerFactory<String, EmojiDto> factory = new ConcurrentKafkaListenerContainerFactory<>();
-        factory.setConsumerFactory(emojiChatConsumerFactory());
+    public ConcurrentKafkaListenerContainerFactory<String, DirectMessageDto> directChatListenerContainerFactory() {
+        ConcurrentKafkaListenerContainerFactory<String, DirectMessageDto> factory = new ConcurrentKafkaListenerContainerFactory<>();
+        factory.setConsumerFactory(directChatConsumerFactory());
         return factory;
     }
 }
