+++ src/backend/chat-service/src/main/java/capstone/chatservice/global/config/kafka/consumer/chat/EmojiChatConsumerConfig.java
@@ -1,6 +1,6 @@
-package capstone.chatservice.global.config.kafka.consumer;
+package capstone.chatservice.global.config.kafka.consumer.chat;
 
-import capstone.chatservice.domain.dm.dto.DirectMessageDto;
+import capstone.chatservice.domain.emoji.dto.EmojiDto;
 import java.util.HashMap;
 import java.util.Map;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
@@ -16,17 +16,16 @@ import org.springframework.kafka.support.serializer.JsonDeserializer;
 
 @EnableKafka
 @Configuration
-public class DirectChatConsumerConfig {
+public class EmojiChatConsumerConfig {
 
     @Value("${spring.kafka.bootstrap-servers}")
     private String bootstrapServers;
 
-    @Value("${spring.kafka.consumer.group-id.direct-chat}")
+    @Value("${spring.kafka.consumer.group-id.emoji-chat}")
     private String groupId;
 
-
     @Bean
-    public Map<String, Object> directChatConsumerConfiguration() {
+    public Map<String, Object> emojiChatConsumerConfiguration() {
         Map<String, Object> config = new HashMap<>();
         config.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapServers);
         config.put(ConsumerConfig.GROUP_ID_CONFIG, groupId);
@@ -34,17 +33,17 @@ public class DirectChatConsumerConfig {
     }
 
     @Bean
-    public ConsumerFactory<String, DirectMessageDto> directChatConsumerFactory() {
+    public ConsumerFactory<String, EmojiDto> emojiChatConsumerFactory() {
         return new DefaultKafkaConsumerFactory<>(
-                directChatConsumerConfiguration(),
+                emojiChatConsumerConfiguration(),
                 new StringDeserializer(),
-                new JsonDeserializer<>(DirectMessageDto.class));
+                new JsonDeserializer<>(EmojiDto.class));
     }
 
     @Bean
-    public ConcurrentKafkaListenerContainerFactory<String, DirectMessageDto> directChatListenerContainerFactory() {
-        ConcurrentKafkaListenerContainerFactory<String, DirectMessageDto> factory = new ConcurrentKafkaListenerContainerFactory<>();
-        factory.setConsumerFactory(directChatConsumerFactory());
+    public ConcurrentKafkaListenerContainerFactory<String, EmojiDto> emojiChatListenerContainerFactory() {
+        ConcurrentKafkaListenerContainerFactory<String, EmojiDto> factory = new ConcurrentKafkaListenerContainerFactory<>();
+        factory.setConsumerFactory(emojiChatConsumerFactory());
         return factory;
     }
 }
