+++ src/backend/chat-service/src/main/java/capstone/chatservice/infra/kafka/consumer/chat/DirectChatConsumer.java
@@ -1,4 +1,4 @@
-package capstone.chatservice.infra.kafka.consumer;
+package capstone.chatservice.infra.kafka.consumer.chat;
 
 import capstone.chatservice.domain.dm.dto.DirectMessageDto;
 import capstone.chatservice.domain.dm.dto.response.DirectMessageCreateResponse;
