+++ src/backend/chat-service/src/main/java/capstone/chatservice/infra/kafka/producer/alarm/AlarmEventProducer.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Component;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class AlarmProducer {
+public class AlarmEventProducer {
 
     @Value("${spring.kafka.topic.dm-alarm-event}")
     private String dmAlarmEventTopic;
