+++ src/backend/chat-service/src/main/java/capstone/chatservice/infra/kafka/producer/chat/ChatEventProducer.java
@@ -1,11 +1,9 @@
-package capstone.chatservice.infra.kafka.producer;
+package capstone.chatservice.infra.kafka.producer.chat;
 
 import capstone.chatservice.domain.dm.dto.DirectMessageDto;
 import capstone.chatservice.domain.emoji.dto.EmojiDto;
 import capstone.chatservice.domain.forum.dto.ForumMessageDto;
 import capstone.chatservice.domain.server.dto.ServerMessageDto;
-import capstone.chatservice.infra.kafka.producer.dto.ConnectionStateEventDto;
-import capstone.chatservice.infra.kafka.producer.dto.ConnectionStateInfo;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -15,7 +13,7 @@ import org.springframework.stereotype.Component;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class KafkaProducer {
+public class ChatEventProducer {
 
     @Value("${spring.kafka.topic.server-chat}")
     private String serverChatTopic;
@@ -29,18 +27,11 @@ public class KafkaProducer {
     @Value("${spring.kafka.topic.forum-chat}")
     private String forumChatTopic;
 
-    @Value("${spring.kafka.topic.connection-state-event}")
-    private String connectionStateEventTopic;
-
-    @Value("${spring.kafka.topic.connection-state-info}")
-    private String connectionStateInfoTopic;
 
     private final KafkaTemplate<String, ServerMessageDto> serverChatKafkaTemplate;
     private final KafkaTemplate<String, DirectMessageDto> direcetChatKafkaTemplate;
     private final KafkaTemplate<String, EmojiDto> emojiChatKafkaTemplate;
     private final KafkaTemplate<String, ForumMessageDto> forumChatKafkaTemplate;
-    private final KafkaTemplate<String, ConnectionStateEventDto> connectionStateEventKafkaTemplate;
-    private final KafkaTemplate<String, ConnectionStateInfo> connectionStateInfoKafkaTemplate;
 
     public void sendToServerChatTopic(ServerMessageDto messageDto) {
         serverChatKafkaTemplate.send(serverChatTopic, messageDto);
@@ -56,13 +47,5 @@ public class KafkaProducer {
 
     public void sendToForumChatTopic(ForumMessageDto forumMessageDto) {
         forumChatKafkaTemplate.send(forumChatTopic, forumMessageDto);
-    }
-
-    public void sendToConnectionStateEventTopic(ConnectionStateEventDto connectionStateEventDto) {
-        connectionStateEventKafkaTemplate.send(connectionStateEventTopic, connectionStateEventDto);
-    }
-
-    public void sendToConnectionStateInfoTopic(ConnectionStateInfo connectionStateInfo) {
-        connectionStateInfoKafkaTemplate.send(connectionStateInfoTopic, connectionStateInfo);
     }
 }
